Move privilege subcommands to privilege component and add directly to iam instead of using custom privilege commands. This commit moves the subcommands file but does not modify it to remove options not needed for tenant.
